Move protocol wrappers to their server
