+++ src/amdinfer/CMakeLists.txt
@@ -30,7 +30,6 @@ add_subdirectory(bindings)
 add_subdirectory(buffers)
 add_subdirectory(clients)
 add_subdirectory(core)
-add_subdirectory(protocol_wrappers)
 add_subdirectory(observation)
 add_subdirectory(servers)
 add_subdirectory(util)
