+++ src/amdinfer/batching/soft.cpp
@@ -32,16 +32,16 @@
 #include "amdinfer/build_options.hpp"    // for AMDINFER_ENABLE_METRICS
 #include "amdinfer/core/exceptions.hpp"  // for invalid_argument
 #include "amdinfer/core/memory_pool/pool.hpp"
-#include "amdinfer/core/parameters.hpp"  // for ParameterMap
+#include "amdinfer/core/parameters.hpp"        // for ParameterMap
+#include "amdinfer/core/protocol_wrapper.hpp"  // for ProtocolWrapper
 #include "amdinfer/core/worker_info.hpp"
 #include "amdinfer/declarations.hpp"         // for ProtocolWrapperPtr
 #include "amdinfer/observation/logging.hpp"  // for Logger, AMDINFER_LOG_DEBUG
 #include "amdinfer/observation/metrics.hpp"  // for Metrics, MetricCounterIDs
 #include "amdinfer/observation/tracing.hpp"  // for Trace
-#include "amdinfer/protocol_wrappers/protocol_wrapper.hpp"  // for ProtocolWrapper
-#include "amdinfer/util/queue.hpp"   // for BlockingConcurrentQueue
-#include "amdinfer/util/thread.hpp"  // for setThreadName
-#include "amdinfer/util/timer.hpp"   // for Timer
+#include "amdinfer/util/queue.hpp"           // for BlockingConcurrentQueue
+#include "amdinfer/util/thread.hpp"          // for setThreadName
+#include "amdinfer/util/timer.hpp"           // for Timer
 
 // default batcher timeout in milliseconds
 constexpr auto kDefaultTimeout = 100;
