+++ src/amdinfer/core/endpoints.cpp
@@ -22,13 +22,13 @@
 #include <cassert>      // for assert
 #include <type_traits>  // for __decay_and_strip<>::__type
 
-#include "amdinfer/batching/batcher.hpp"  // for Batcher
-#include "amdinfer/build_options.hpp"     // for kMaxModelNameSize
-#include "amdinfer/core/exceptions.hpp"   // for invalid_argument
-#include "amdinfer/core/parameters.hpp"   // for ParameterMap
-#include "amdinfer/core/worker_info.hpp"  // for WorkerInfo
-#include "amdinfer/protocol_wrappers/protocol_wrapper.hpp"  // IWYU pragma: keep
-#include "amdinfer/util/thread.hpp"                         // for setThreadName
+#include "amdinfer/batching/batcher.hpp"       // for Batcher
+#include "amdinfer/build_options.hpp"          // for kMaxModelNameSize
+#include "amdinfer/core/exceptions.hpp"        // for invalid_argument
+#include "amdinfer/core/parameters.hpp"        // for ParameterMap
+#include "amdinfer/core/protocol_wrapper.hpp"  // IWYU pragma: keep
+#include "amdinfer/core/worker_info.hpp"       // for WorkerInfo
+#include "amdinfer/util/thread.hpp"            // for setThreadName
 
 namespace amdinfer {
