+++ src/amdinfer/core/protocol_wrapper.hpp
