+++ src/amdinfer/core/worker_info.cpp
@@ -33,10 +33,10 @@
 
 #include "amdinfer/batching/batcher.hpp"  // for Batcher, BatcherStatus, Bat...
 #include "amdinfer/core/exceptions.hpp"   // for invalid_argument, external_...
-#include "amdinfer/core/memory_pool/pool.hpp"               // for MemoryPool
-#include "amdinfer/core/parameters.hpp"                     // for ParameterMap
-#include "amdinfer/core/predict_api.hpp"                    // for ModelMetadata
-#include "amdinfer/protocol_wrappers/protocol_wrapper.hpp"  // IWYU pragma: keep
+#include "amdinfer/core/memory_pool/pool.hpp"  // for MemoryPool
+#include "amdinfer/core/parameters.hpp"        // for ParameterMap
+#include "amdinfer/core/predict_api.hpp"       // for ModelMetadata
+#include "amdinfer/core/protocol_wrapper.hpp"  // IWYU pragma: keep
 #include "amdinfer/workers/worker.hpp"  // for Worker, WorkerStatus, Worke...
 
 namespace amdinfer {
