+++ tests/unit/batching/test_soft.cpp
@@ -18,10 +18,10 @@
 #include "amdinfer/batching/soft.hpp"          // for SoftBatcher
 #include "amdinfer/build_options.hpp"          // for AMDINFER_ENABLE_LOGGING
 #include "amdinfer/core/memory_pool/pool.hpp"  // for MemoryPool
+#include "amdinfer/core/protocol_wrapper.hpp"  // IWYU pragma: keep
 #include "amdinfer/core/worker_info.hpp"       // for WorkerInfo
 #include "amdinfer/observation/logging.hpp"  // for initLogger, LogLevel, Log...
-#include "amdinfer/protocol_wrappers/protocol_wrapper.hpp"  // IWYU pragma: keep
-#include "gtest/gtest.h"  // for Test, SuiteApiResolver, TEST
+#include "gtest/gtest.h"                     // for Test, SuiteApiResolver, TEST
 
 namespace amdinfer {
